Orders endpoints

+++ src/products/products.controller.ts
@@ -1,7 +1,10 @@
-import { Controller, Get, Delete, Post, Body } from '@nestjs/common';
+import { Controller, Get, Delete, Post, Body, Put } from '@nestjs/common';
 import { ProductsService } from './products.service';
 import { Param } from '@nestjs/common';
 import { CreateProductDTO } from './dtos/create-product.dto';
+import { ParseUUIDPipe } from '@nestjs/common';
+import { UpdateProductDTO } from './dtos/update-product.dto'
+import { NotFoundException } from '@nestjs/common/exceptions';
 
 @Controller('products')
 export class ProductsController {
@@ -15,12 +18,16 @@ export class ProductsController {
     }
 
     @Get('/:id')
-    public getById(@Param('id') id: string) {
-        return this.productsService.getById(id);
+    getById(@Param('id', new ParseUUIDPipe()) id: string) {
+        const prod = this.productsService.getById(id);
+        if (!prod) throw new NotFoundException('Product not found');
+        return prod;
     }
 
     @Delete('/:id')
-    deleteById(@Param('id') id: string) {
+    deleteById(@Param('id', new ParseUUIDPipe()) id: string) {
+        if (!this.productsService.getById(id))
+            throw new NotFoundException('Product not found');
         this.productsService.deleteById(id);
         return { success: true };
     }
@@ -28,6 +35,18 @@ export class ProductsController {
     @Post('/')
     create(@Body() productData: CreateProductDTO) {
         return this.productsService.create(productData);
+    }
+
+    @Put('/:id')
+    update(
+        @Param('id', new ParseUUIDPipe()) id: string,
+        @Body() productData: UpdateProductDTO,
+    ) {
+        if (!this.productsService.getById(id))
+            throw new NotFoundException('Product not found');
+
+        this.productsService.updateById(id, productData);
+        return { success: true };
     }
 }
